remove obsolete code, split each dialog/browser in one folder
There has been a lot of changed in the UI in the past weeks. So there was quite a lot of code that wasn't used anymore. In addition, the concept in which folder each belongs changed. Now, the files are grouped by window/dialog instead of by file type (e.g. .css, .html). This ensures high cohesion.